add doc + move
