+++ src/Client/WebToElasticms/Extract/Extractor.php
@@ -7,8 +7,8 @@ namespace App\Client\WebToElasticms\Extract;
 use App\Client\HttpClient\CacheManager;
 use App\Client\WebToElasticms\Config\Computer;
 use App\Client\WebToElasticms\Config\ConfigManager;
-use App\Client\WebToElasticms\Config\DocumentResources;
 use App\Client\WebToElasticms\Config\WebResource;
+use App\Client\WebToElasticms\Helper\DocumentResources;
 use App\Client\WebToElasticms\Helper\ExpressionData;
 use App\Client\WebToElasticms\Helper\Url;
 use App\Client\WebToElasticms\Rapport\Rapport;
